fix bugs, add used futures saving
Former-commit-id: 31f59251969ecd426117493c942f10fb872aeeab

+++ future_r=3_10/10/makeHistoricalData.py
@@ -368,7 +368,8 @@ def makeHistoricalData(h, r, test_size, target, feature_selection, spatial_mode,
         allData = pd.merge(independantOfTimeData, timeDeapandantData, on='county_fips')
     else:
         allData = timeDeapandantData
-
+    allData['death'] = allData['death'].astype(float)
+    allData['confirmed'] = allData['confirmed'].astype(float)
     allData = allData.sort_values(by=['date', 'county_fips'])
     allData = allData.reset_index(drop=True)
